requirements.txt
pandas==0.23.4

yaml-1.3==0.1.0

requests==2.22.0

urllib3==1.25.3

numpy==1.18.2

PyYAML==5.2